backend/requirements.txt
uvicorn[standard]
fastapi
sqlmodel
psycopg2-binary